better for testing

+++ Projects/euler_2d/main.cpp
@@ -24,7 +24,7 @@ using namespace PhysBAM;
 
 int main(int argc,char* argv[])
 {
-    typedef float T;
+    typedef double T;
     typedef float RW;
     typedef VECTOR<T,2> TV;
